+++ tests/integration/test_models/test_full_battery_models/test_lead_acid/test_composite.py
@@ -87,7 +87,7 @@ class TestLeadAcidCompositeSurfaceForm(unittest.TestCase):
         modeltest.test_all()  # solver=pybamm.CasadiSolver())
 
     # def test_thermal(self):
-    #     options = {"thermal": "x-lumped"}
+    #     options = {"thermal": "lumped"}
     #     model = pybamm.lead_acid.Composite(options)
     #     modeltest = tests.StandardModelTest(model)
     #     modeltest.test_all()